requirements.txt

streamlit>=1.36
pandas>=2.0
jsonschema>=4.21
python-dotenv>=1.0
